+++ CHANGELOG.md
@@ -12,8 +12,11 @@
 - The internals of `Tokens` are no longer public. `Tokens` can no longer be constructed by user code, and is now only returned by the `Serializer`.
 - Comparison with a `Tokens` can now be done with any type that implements `IntoIterator<Item = &Token>`.
 - `de::Builder::build()` now only requires `&self` instead of `&mut self`.
+- `Error::ExpectedToken` variant has been replaced by `Error::ExpectedSeqEnd`, `Error::ExpectedTupleEnd`, ``Error::ExpectedTupleStructEnd`, `Error::ExpectedTupleVariantEnd`, `Error::ExpectedMapEnd`, `Error::ExpectedStructEnd`, and `Error::ExpectedStructVariantEnd` variants.
 ### Removed
 - `From<Token>` implementation for `serde::de::Unexpected`.
+- `PartialEq` implementation for `Token`.
+- Dependency on `hashbrown` crate.
 
 ## 0.6.0 - 2023-11-19
 ### Changed